Implemented caching in Catalogue and DataExport repositories
